+++ lib/src/classes/legacy_uuid.dart
@@ -4,23 +4,31 @@ import 'package:uuid/uuid.dart';
 
 import '../types/bson_binary.dart';
 
-// To create BinData values corresponding to the various driver encodings use:
-//      var s = "{00112233-4455-6677-8899-aabbccddeeff}";
-//      var uuid = UUID(s); // new Standard encoding
-//      var juuid = JUUID(s); // JavaLegacy encoding
-//      var csuuid = CSUUID(s); // CSharpLegacy encoding
-//      var pyuuid = PYUUID(s); // PythonLegacy encoding
-// To convert the various BinData values back to human readable UUIDs use:
-//      uuid.toUUID()     => 'UUID("00112233-4455-6677-8899-aabbccddeeff")'
-//      juuid.ToJUUID()   => 'JUUID("00112233-4455-6677-8899-aabbccddeeff")'
-//      csuuid.ToCSUUID() => 'CSUUID("00112233-4455-6677-8899-aabbccddeeff")'
-//      pyuuid.ToPYUUID() => 'PYUUID("00112233-4455-6677-8899-aabbccddeeff")'
-// With any of the UUID variants you can use toHexUUID to echo the raw BinData with subtype and hex string:
-//      uuid.toHexUUID()   => 'HexData(4, "00112233-4455-6677-8899-aabbccddeeff")'
-//      juuid.toHexUUID()  => 'HexData(3, "77665544-3322-1100-ffee-ddccbbaa9988")'
-//      csuuid.toHexUUID() => 'HexData(3, "33221100-5544-7766-8899-aabbccddeeff")'
-//      pyuuid.toHexUUID() => 'HexData(3, "00112233-4455-6677-8899-aabbccddeeff")'
-
+/// This class is a convenient way for managing Binary subtype 3 Bson
+/// types (Uuid Old).
+/// The class holds the content in the format that will be stored in the
+/// data base.
+/// You can directly create the object with .fromHexString(), or
+/// use the format converters for each flavour that have been used
+/// As per today, the flavours I'm aware of are: Java legacy, C# legacy
+/// and Python legacy.
+/// The storing algorithn are the following (found in internet,
+/// please confirm me):
+/// Standard (subtype 4, not covered from this class):
+/// -- From value "00112233-4455-6677-8899-aabbccddeeff"
+/// -- To value "00112233-4455-6677-8899-aabbccddeeff"
+/// Java:
+/// -- From value "00112233-4455-6677-8899-aabbccddeeff"
+/// -- To value "77665544-3322-1100-ffee-ddccbbaa9988"
+/// C#:
+/// -- From value "00112233-4455-6677-8899-aabbccddeeff"
+/// -- To value "33221100-5544-7766-8899-aabbccddeeff"
+/// Python:
+/// -- From value "00112233-4455-6677-8899-aabbccddeeff"
+/// -- To value "00112233-4455-6677-8899-aabbccddeeff"
+///
+/// Once you will read these values, you will have to used the correspondent
+/// flavour method to get back the originl value (ex. javaLegacy())
 class LegacyUuid {
   /// The ByteList Expected is in the desired legacy format to store
   LegacyUuid(this.content);